+++ src/components/Filters/FilterFilters.js
@@ -1,5 +1,5 @@
 import PropTypes from 'prop-types';
-import { useState, useEffect } from 'react';
+import { useState } from 'react';
 import { FormattedMessage } from 'react-intl';
 
 import {
@@ -9,6 +9,7 @@ import {
 } from '@folio/stripes/components';
 import { CheckboxFilter } from '@folio/stripes/smart-components';
 
+import useUpdatedFilters from '../hooks/useUpdatedFilters';
 import filterConfig from './filterConfigData';
 
 const FilterFilters = ({
@@ -19,35 +20,13 @@ const FilterFilters = ({
     type: [],
   });
 
-  useEffect(() => {
-    const newState = {};
-    const arr = [];
-
-    filterConfig.forEach(filter => {
-      const newValues = [];
-      let values = {};
-      values = filter.values;
-
-      values.forEach((key) => {
-        let newValue = {};
-        newValue = {
-          'value': key.cql,
-          'label': key.name,
-        };
-        newValues.push(newValue);
-      });
-
-      arr[filter.name] = newValues;
-
-      if (filterState[filter.name] && arr[filter.name].length !== filterState[filter.name].length) {
-        newState[filter.name] = arr[filter.name];
-      }
-    });
-
-    if (Object.keys(newState).length) {
-      setFilterState((prevState) => ({ ...prevState, ...newState }));
-    }
-  }, [filterState]);
+  useUpdatedFilters({
+    dynamicKey: '',
+    filterConfig,
+    filterData: [],
+    filterState,
+    setFilterState,
+  });
 
   const renderCheckboxFilter = (key, props) => {
     const groupFilters = activeFilters[key] || [];
