+++ src/components/MetadataCollections/CollectionFilters.js
@@ -1,5 +1,5 @@
 import PropTypes from 'prop-types';
-import { useState, useEffect } from 'react';
+import { useState } from 'react';
 import { FormattedMessage } from 'react-intl';
 
 import {
@@ -10,6 +10,7 @@ import {
 } from '@folio/stripes/components';
 import { CheckboxFilter } from '@folio/stripes/smart-components';
 
+import useUpdatedFilters from '../hooks/useUpdatedFilters';
 import filterConfig from './filterConfigData';
 
 const CollectionFilters = ({
@@ -30,41 +31,13 @@ const CollectionFilters = ({
     mdSource: [],
   });
 
-  useEffect(() => {
-    const newState = {};
-    const arr = [];
-
-    filterConfig.forEach(filter => {
-      const newValues = [];
-      let values = {};
-      if (filter === 'mdSource') {
-        // get filter values from okapi
-        values = filterData[filter] || [];
-      } else {
-        // get filte values from filterConfig
-        values = filter.values;
-      }
-
-      values.forEach((key) => {
-        let newValue = {};
-        newValue = {
-          'value': key.cql,
-          'label': key.name,
-        };
-        newValues.push(newValue);
-      });
-
-      arr[filter.name] = newValues;
-
-      if (filterState[filter.name] && arr[filter.name].length !== filterState[filter.name].length) {
-        newState[filter.name] = arr[filter.name];
-      }
-    });
-
-    if (Object.keys(newState).length) {
-      setFilterState((prevState) => ({ ...prevState, ...newState }));
-    }
-  }, [filterData, filterState]);
+  useUpdatedFilters({
+    dynamicKey: 'mdSource',
+    filterConfig,
+    filterData,
+    filterState,
+    setFilterState,
+  });
 
   const renderCheckboxFilter = (key) => {
     const groupFilters = activeFilters[key] || [];
